Summary of Fixes 1. Event Creation Flow - min_participants Stage - File: /home/zorvan/Work/projects/Zwischen/telegram-bot/bot/commands/flow/__init__.py - Issue: Bot stopped after asking for minimum number of people because there was no message handler for the min_participants stage - Fix: Added handler at line 1145 that processes user's text input as minimum participant count 2. Constraint Creation - Database Type Mismatch - Files:   - /home/zorvan/Work/projects/Zwischen/telegram-bot/db/schema.sql   - /home/zorvan/Work/projects/Zwischen/telegram-bot/db/models.py   - /home/zorvan/Work/projects/Zwischen/telegram-bot/db/migrations/fix_constraint_types.sql (new) - Issue: Database had INTEGER columns for user_id and target_user_id in the constraints table, but users.user_id is BIGINT, causing type mismatch and IntegrityError - Fix: Updated both schema.sql and models.py to use BIGINT/BigInteger consistently, and applied migration to fix existing database 3. Constraint Creation - Duplicate Check - File: /home/zorvan/Work/projects/Zwischen/telegram-bot/bot/commands/event_details.py - Issue: No check for existing availability constraints before saving - Fix: Added check at line 928-939 to verify constraint doesn't already exist before attempting to save

+++ telegram-bot/db/models.py
@@ -55,7 +55,7 @@ class Group(Base):
 
     __tablename__ = "groups"
 
-    group_id = Column(Integer, primary_key=True)
+    group_id = Column(BigInteger, primary_key=True)
     telegram_group_id = Column(BigInteger, unique=True, nullable=False)
     group_name = Column(String(255))
     group_type = Column(String(50), default="casual")
@@ -72,7 +72,7 @@ class Event(Base):
 
     event_id = Column(BigInteger, primary_key=True)
     group_id = Column(
-        Integer, ForeignKey("groups.group_id", ondelete="CASCADE"), nullable=False
+        BigInteger, ForeignKey("groups.group_id", ondelete="CASCADE"), nullable=False
     )
     event_type = Column(String(100), nullable=False)
     description = Column(Text)
